% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_survey}
\alias{run_survey}
\title{Run the entire pipeline}
\usage{
run_survey(parameter_path = "params.json")
}
\arguments{
\item{parameter_path}{path to the parameters file}
}
\value{
none
}
\description{
See the README.md file in the source directory for more information about how to run
this function.
}
